Update JUnit version

+++ src/main/java/ru/gb/testing/Calculator.java
@@ -1,8 +1,8 @@
 package ru.gb.testing;
 
 public class Calculator {
-    public int add(int x1, int x2) {
-        return x1 + x2;
+    public long add(int x1, int x2) {
+        return (long) x1 + x2;
     }
 
     public int sub(int x1, int x2) {
